perf: seventh batch - dispatch fix, Python stack, BatchNorm eval, depthwise
- dispatch(): use unwrap_or_else instead of expect+format! to avoid
  string allocation on every kernel dispatch (was ~100 allocs/step)
- Python stack(): use np.stack instead of element-by-element Python loop
  (was O(batch*features) Python iterations per batch)
- Python numpy(): pass dtype=np.float32 upfront to avoid type-inference copy
- Python sum(dim=None)/mean(dim=None): flatten+single dispatch instead of
  O(dims) Rust round-trips
- BatchNorm1d: hold RwLock read guard in eval mode to avoid cloning
  running_mean/running_var tensors every forward
- Attention: pre-compute scale as f32 field, use mul_scalar() instead of
  Tensor::from_scalar(scale) per forward
- Conv2d: pre-allocate default_bias tensor, use reference instead of
  clone-or-allocate pattern
- Tensor: add non-in-place mul_scalar() method avoiding scalar tensor alloc
- depthwise_conv2d: use raw pointers instead of .to_vec() copies (was
  copying entire weight+input tensors per call)

+++ fastnn/__init__.py
@@ -66,7 +66,7 @@ def _patch_numpy(tensor_cls):
     def _new_numpy(self):
         data = _original_numpy(self)
         shape = self.shape
-        return np.array(data).reshape(shape)
+        return np.array(data, dtype=np.float32).reshape(shape)
 
     tensor_cls.numpy = _new_numpy
